+++ OOS/src/app/admin/orders/orders.component.html
@@ -6,8 +6,8 @@
   <tr>
     <th>Order Email</th>
     <th>User ID</th>
-    <th>Address</th>
     <th>Total</th>
+    <th>Action</th>
   </tr>
   <tr class="collection-item" *ngFor="let order of orders; let i = index">
     <td>
